WIP on player turn
